requirements.txt
pywebview
mss
numpy
pypiwin32
Pillow
